Wow. I did everything.
Okay, not everything. I'm trying to load most everything as a custom
component. The GameBoard is a custom component, which in turn loads the
ScoreBoard which is a custom component, which in turn loads the
BombCounter, which is also a custom component; the BombCounter loads
itself and then we head on back.

Hopefully, this custom component business will make things really
simple. I'll have components that house their own capabilities AND—
get this—will be visible in SceneBuilder. That's my next step, I need
to make a jar out of what I have and then load that in to SceneBuilder
in order to import the custom components and by so doing, be able to
edit each one in scene builder.